Reorganize some binder related code.
