+++ src/tools/wix/Bind/ChainPackageInfo.cs
@@ -11,7 +11,7 @@
 // </summary>
 //-------------------------------------------------------------------------------------------------
 
-namespace WixToolset
+namespace WixToolset.Bind
 {
     using System;
     using System.Collections;
